+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement Copy and Eq for EspTwaiError (#540)
 - Add LEDC hardware fade support
 - Added support for multicore async GPIO (#542)
+- Add initial support for MCPWM in ESP32-H2 (#544)
 
 ### Fixed